+++ qcelemental/models/v1/procedures.py
@@ -1,5 +1,11 @@
 from enum import Enum
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Union
+
+try:
+    from typing import Literal
+except ImportError:
+    # remove when minimum py38
+    from typing_extensions import Literal
 
 try:
     from typing import Literal
@@ -16,6 +22,7 @@ from .common_models import (
     DriverEnum,
     Model,
     Provenance,
+    check_convertible_version,
     qcschema_input_default,
     qcschema_optimization_input_default,
     qcschema_optimization_output_default,
@@ -59,7 +66,7 @@ class QCInputSpecification(ProtoModel):
     """
 
     schema_name: Literal["qcschema_input"] = Field(qcschema_input_default)
-    schema_version: int = 1
+    schema_version: int = 1  # TODO
 
     driver: DriverEnum = Field(DriverEnum.gradient, description=str(DriverEnum.__doc__))
     model: Model = Field(..., description=str(Model.__doc__))
@@ -79,7 +86,7 @@ class OptimizationInput(ProtoModel):
     schema_name: constr(  # type: ignore
         strip_whitespace=True, regex=qcschema_optimization_input_default
     ) = qcschema_optimization_input_default
-    schema_version: int = 1
+    schema_version: Literal[1] = 1
 
     keywords: Dict[str, Any] = Field({}, description="The optimization specific keywords to be used.")
     extras: Dict[str, Any] = Field({}, description="Extra fields that are not part of the schema.")
@@ -96,6 +103,25 @@ class OptimizationInput(ProtoModel):
             ("molecule_hash", self.initial_molecule.get_hash()[:7]),
         ]
 
+    @validator("schema_version", pre=True)
+    def _version_stamp(cls, v):
+        return 1
+
+    def convert_v(
+        self, version: int
+    ) -> Union["qcelemental.models.v1.OptimizationInput", "qcelemental.models.v2.OptimizationInput"]:
+        """Convert to instance of particular QCSchema version."""
+        import qcelemental as qcel
+
+        if check_convertible_version(version, error="OptimizationInput") == "self":
+            return self
+
+        dself = self.dict()
+        if version == 2:
+            self_vN = qcel.models.v2.OptimizationInput(**dself)
+
+        return self_vN
+
 
 class OptimizationResult(OptimizationInput):
     """QCSchema results model for geometry optimization."""
@@ -103,6 +129,7 @@ class OptimizationResult(OptimizationInput):
     schema_name: constr(  # type: ignore
         strip_whitespace=True, regex=qcschema_optimization_output_default
     ) = qcschema_optimization_output_default
+    schema_version: Literal[1] = 1
 
     final_molecule: Optional[Molecule] = Field(..., description="The final molecule of the geometry optimization.")
     trajectory: List[AtomicResult] = Field(
@@ -141,6 +168,25 @@ class OptimizationResult(OptimizationInput):
 
         return v
 
+    @validator("schema_version", pre=True)
+    def _version_stamp(cls, v):
+        return 1
+
+    def convert_v(
+        self, version: int
+    ) -> Union["qcelemental.models.v1.OptimizationResult", "qcelemental.models.v2.OptimizationResult"]:
+        """Convert to instance of particular QCSchema version."""
+        import qcelemental as qcel
+
+        if check_convertible_version(version, error="OptimizationResult") == "self":
+            return self
+
+        dself = self.dict()
+        if version == 2:
+            self_vN = qcel.models.v2.OptimizationResult(**dself)
+
+        return self_vN
+
 
 class OptimizationSpecification(ProtoModel):
     """
@@ -153,7 +199,7 @@ class OptimizationSpecification(ProtoModel):
     """
 
     schema_name: constr(strip_whitespace=True, regex="qcschema_optimization_specification") = "qcschema_optimization_specification"  # type: ignore
-    schema_version: int = 1
+    schema_version: int = 1  # TODO
 
     procedure: str = Field(..., description="Optimization procedure to run the optimization with.")
     keywords: Dict[str, Any] = Field({}, description="The optimization specific keywords to be used.")
@@ -210,7 +256,7 @@ class TorsionDriveInput(ProtoModel):
     """
 
     schema_name: constr(strip_whitespace=True, regex=qcschema_torsion_drive_input_default) = qcschema_torsion_drive_input_default  # type: ignore
-    schema_version: int = 1
+    schema_version: Literal[1] = 1
 
     keywords: TDKeywords = Field(..., description="The torsion drive specific keywords to be used.")
     extras: Dict[str, Any] = Field({}, description="Extra fields that are not part of the schema.")
@@ -231,6 +277,25 @@ class TorsionDriveInput(ProtoModel):
         assert value.driver == DriverEnum.gradient, "driver must be set to gradient"
         return value
 
+    @validator("schema_version", pre=True)
+    def _version_stamp(cls, v):
+        return 1
+
+    def convert_v(
+        self, version: int
+    ) -> Union["qcelemental.models.v1.TorsionDriveInput", "qcelemental.models.v2.TorsionDriveInput"]:
+        """Convert to instance of particular QCSchema version."""
+        import qcelemental as qcel
+
+        if check_convertible_version(version, error="TorsionDriveInput") == "self":
+            return self
+
+        dself = self.dict()
+        if version == 2:
+            self_vN = qcel.models.v2.TorsionDriveInput(**dself)
+
+        return self_vN
+
 
 class TorsionDriveResult(TorsionDriveInput):
     """Results from running a torsion drive.
@@ -241,7 +306,7 @@ class TorsionDriveResult(TorsionDriveInput):
     """
 
     schema_name: constr(strip_whitespace=True, regex=qcschema_torsion_drive_output_default) = qcschema_torsion_drive_output_default  # type: ignore
-    schema_version: int = 1
+    schema_version: Literal[1] = 1
 
     final_energies: Dict[str, float] = Field(
         ..., description="The final energy at each angle of the TorsionDrive scan."
@@ -263,6 +328,25 @@ class TorsionDriveResult(TorsionDriveInput):
     )
     error: Optional[ComputeError] = Field(None, description=str(ComputeError.__doc__))
     provenance: Provenance = Field(..., description=str(Provenance.__doc__))
+
+    @validator("schema_version", pre=True)
+    def _version_stamp(cls, v):
+        return 1
+
+    def convert_v(
+        self, version: int
+    ) -> Union["qcelemental.models.v1.TorsionDriveResult", "qcelemental.models.v2.TorsionDriveResult"]:
+        """Convert to instance of particular QCSchema version."""
+        import qcelemental as qcel
+
+        if check_convertible_version(version, error="TorsionDriveResult") == "self":
+            return self
+
+        dself = self.dict()
+        if version == 2:
+            self_vN = qcel.models.v2.TorsionDriveResult(**dself)
+
+        return self_vN
 
 
 def Optimization(*args, **kwargs):